requirements.txt
certifi==2022.12.7
click==8.1.3
Flask==2.2.2
googletrans==4.0.0rc1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.23.4
openai==0.26.5
Pillow==9.4.0
psycopg2==2.9.5
requests==2.28.2
scipy==1.10.0
six==1.16.0
sumy==0.11.0
transformers==4.26.0
Werkzeug==2.2.2
wordcloud==1.8.2.2
